Canonical link: https://commits.webkit.org/152923.281@safari-600.1.4.13-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-600.1.4.13-branch@177480 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/ios/WebVideoFullscreenInterfaceAVKit.mm
@@ -684,7 +684,7 @@ void WebVideoFullscreenInterfaceAVKit::setAudioMediaSelectionOptions(const Vecto
     dispatch_async(dispatch_get_main_queue(), [webOptions, strongThis, selectedIndex] {
         strongThis->playerController().audioMediaSelectionOptions = webOptions.get();
         if (selectedIndex < [webOptions count])
-            strongThis->playerController().currentAudioMediaSelectionOption = [webOptions objectAtIndex:selectedIndex];
+            strongThis->playerController().currentAudioMediaSelectionOption = [webOptions objectAtIndex:static_cast<NSUInteger>(selectedIndex)];
     });
 }
 
@@ -696,7 +696,7 @@ void WebVideoFullscreenInterfaceAVKit::setLegibleMediaSelectionOptions(const Vec
     dispatch_async(dispatch_get_main_queue(), [webOptions, strongThis, selectedIndex] {
         strongThis->playerController().legibleMediaSelectionOptions = webOptions.get();
         if (selectedIndex < [webOptions count])
-            strongThis->playerController().currentLegibleMediaSelectionOption = [webOptions objectAtIndex:selectedIndex];
+            strongThis->playerController().currentLegibleMediaSelectionOption = [webOptions objectAtIndex:static_cast<NSUInteger>(selectedIndex)];
     });
 }
